refactor(editor): Rename canvasTidyUpQueue to pendingCanvasTidyUp (no-changelog)
It holds a single overwritable pending request, not a FIFO queue — the
name should match the single-slot semantics its own exports already use
(PendingCanvasTidyUp, usePendingCanvasTidyUp, consumePendingCanvasTidyUp).
